RAG-backend/pyproject.toml
[project]
name = "bk-python"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.121.0",
    "openai-agents>=0.5.0",
    "python-dotenv>=1.2.1",
]
